Add support for supplying multiple label files
- Fixes #49

This change introduces support for any number of `--labels` arguments to
be supplied. The behavior is the same as before, with regards to http(s)
urls or local paths, and after they are all collected, they will be
merged. Any items with a duplicate name and different property values
will throw an error. Otherwise they will be merged and then applied.