Record all current label matches as resolved
There are two scenarios where multiple current labels can match a configured label:

- The configured label exists in the current labels and an alias exists in the current labels
- Multiple aliases of the configured label exist in the current labels

As a partial workaround for the lack of a label merge capability in GitHub Label Sync, only the first matching label is
processed for the creation of a diff entry. This approach was mistakenly applied also to adding matches to the list of
"resolved" labels. That caused the tool to think that any additional matching labels did not match any label
configuration, which resulted in "added" entries being created for them, and their incorrect deletion in the default
mode where additional labels are not allowed.